+++ import/clipboard/clipboard_linux_bsd.go
@@ -1,7 +1,7 @@
 // +build linux freebsd netbsd openbsd dragonfly
 
 /*
- * Copyright (C) 2014-2015 Vi Grey. All rights reserved.
+ * Copyright (C) 2014-2015 Dev Brush Technology. All rights reserved.
  * Copyright (C) 2013 Ato Araki. All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without